move button component to bright mirror common folder
